Moving capacity engine to overlay.
